requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.5
dnspython==2.1.0
Flask==2.0.2
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
nltk==3.7
numpy==1.21.4
packaging==21.3
pandas==1.3.4
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pymongo==4.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2021.3
regex==2022.6.2
requests==2.26.0
scikit-learn==1.0.1
scipy==1.7.3
six==1.16.0
sklearn==0.0
spotipy==2.19.0
threadpoolctl==3.0.0
tomli==2.0.1
tqdm==4.64.0
urllib3==1.26.7
Werkzeug==2.0.2
wincertstore==0.2
xgboost==1.5.1
